requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1
supabase==2.18.1
pydantic==2.11.7
pydantic-settings==2.0.3
email-validator==2.2.0
beautifulsoup4==4.12.2
lxml==5.1.0
httpx==0.27.0
